refactor: use unique sql template set
