+++ src/common/meta/src/kv_backend/rds.rs
@@ -33,252 +33,11 @@ use crate::rpc::store::{
 };
 use crate::rpc::KeyValue;
 
-type PgClient = deadpool::managed::Object<deadpool_postgres::Manager>;
+mod postgres;
 
-enum PgQueryExecutor<'a> {
-    Client(PgClient),
-    Transaction(deadpool_postgres::Transaction<'a>),
-}
+pub use postgres::PgStore;
 
-impl PgQueryExecutor<'_> {
-    async fn query(
-        &self,
-        query: &str,
-        params: &[&(dyn ToSql + Sync)],
-    ) -> Result<Vec<tokio_postgres::Row>> {
-        match self {
-            PgQueryExecutor::Client(client) => {
-                let stmt = client
-                    .prepare_cached(query)
-                    .await
-                    .context(PostgresExecutionSnafu { sql: query })?;
-                client
-                    .query(&stmt, params)
-                    .await
-                    .context(PostgresExecutionSnafu { sql: query })
-            }
-            PgQueryExecutor::Transaction(txn) => {
-                let stmt = txn
-                    .prepare_cached(query)
-                    .await
-                    .context(PostgresExecutionSnafu { sql: query })?;
-                txn.query(&stmt, params)
-                    .await
-                    .context(PostgresExecutionSnafu { sql: query })
-            }
-        }
-    }
-
-    async fn commit(self) -> Result<()> {
-        match self {
-            PgQueryExecutor::Client(_) => Ok(()),
-            PgQueryExecutor::Transaction(txn) => {
-                txn.commit().await.context(PostgresTransactionSnafu {
-                    operation: "commit".to_string(),
-                })
-            }
-        }
-    }
-}
-
-const PG_STORE_TXN_RETRY_COUNT: usize = 3;
-
-/// Posgres backend store for metasrv
-pub struct PgStore<T> {
-    pool: Pool,
-    max_txn_ops: usize,
-    sql_template_set: SqlTemplateSet<T>,
-    txn_retry_count: usize,
-}
-
-const EMPTY: &[u8] = &[0];
 const RDS_STORE_TXN_RETRY_COUNT: usize = 3;
-
-/// Type of range template.
-#[derive(Debug, Clone, Copy)]
-enum RangeTemplateType {
-    Point,
-    Range,
-    Full,
-    LeftBounded,
-    Prefix,
-}
-
-/// Builds params for the given range template type.
-impl RangeTemplateType {
-    fn build_params(&self, mut key: Vec<u8>, range_end: Vec<u8>) -> Vec<Vec<u8>> {
-        match self {
-            RangeTemplateType::Point => vec![key],
-            RangeTemplateType::Range => vec![key, range_end],
-            RangeTemplateType::Full => vec![],
-            RangeTemplateType::LeftBounded => vec![key],
-            RangeTemplateType::Prefix => {
-                key.push(b'%');
-                vec![key]
-            }
-        }
-    }
-}
-
-/// Templates for range request.
-#[derive(Debug, Clone)]
-struct RangeTemplate {
-    point: String,
-    range: String,
-    full: String,
-    left_bounded: String,
-    prefix: String,
-}
-
-impl RangeTemplate {
-    /// Gets the template for the given type.
-    fn get(&self, typ: RangeTemplateType) -> &str {
-        match typ {
-            RangeTemplateType::Point => &self.point,
-            RangeTemplateType::Range => &self.range,
-            RangeTemplateType::Full => &self.full,
-            RangeTemplateType::LeftBounded => &self.left_bounded,
-            RangeTemplateType::Prefix => &self.prefix,
-        }
-    }
-
-    /// Adds limit to the template.
-    fn with_limit(template: &str, limit: i64) -> String {
-        if limit == 0 {
-            return format!("{};", template);
-        }
-        format!("{} LIMIT {};", template, limit)
-    }
-}
-
-fn is_prefix_range(start: &[u8], end: &[u8]) -> bool {
-    if start.len() != end.len() {
-        return false;
-    }
-    let l = start.len();
-    let same_prefix = start[0..l - 1] == end[0..l - 1];
-    if let (Some(rhs), Some(lhs)) = (start.last(), end.last()) {
-        return same_prefix && (*rhs + 1) == *lhs;
-    }
-    false
-}
-
-/// Determine the template type for range request.
-fn range_template(key: &[u8], range_end: &[u8]) -> RangeTemplateType {
-    match (key, range_end) {
-        (_, &[]) => RangeTemplateType::Point,
-        (EMPTY, EMPTY) => RangeTemplateType::Full,
-        (_, EMPTY) => RangeTemplateType::LeftBounded,
-        (start, end) => {
-            if is_prefix_range(start, end) {
-                RangeTemplateType::Prefix
-            } else {
-                RangeTemplateType::Range
-            }
-        }
-    }
-}
-
-/// Generate in placeholders for sql.
-fn generate_in_placeholders(from: usize, to: usize) -> Vec<String> {
-    (from..=to).map(|i| format!("${}", i)).collect()
-}
-
-/// Factory for building sql templates.
-struct SqlTemplateFactory<'a> {
-    table_name: &'a str,
-}
-
-impl<'a> SqlTemplateFactory<'a> {
-    /// Creates a new [`SqlTemplateFactory`] with the given table name.
-    fn new(table_name: &'a str) -> Self {
-        Self { table_name }
-    }
-
-    /// Builds the template set for the given table name.
-    fn build(&self) -> SqlTemplateSet {
-        let table_name = self.table_name;
-        SqlTemplateSet {
-            table_name: table_name.to_string(),
-            create_table_statement: format!(
-                "CREATE TABLE IF NOT EXISTS {table_name}(k bytea PRIMARY KEY, v bytea)",
-            ),
-            range_template: RangeTemplate {
-                point: format!("SELECT k, v FROM {table_name} WHERE k = $1"),
-                range: format!("SELECT k, v FROM {table_name} WHERE k >= $1 AND k < $2 ORDER BY k"),
-                full: format!("SELECT k, v FROM {table_name} $1 ORDER BY k"),
-                left_bounded: format!("SELECT k, v FROM {table_name} WHERE k >= $1 ORDER BY k"),
-                prefix: format!("SELECT k, v FROM {table_name} WHERE k LIKE $1 ORDER BY k"),
-            },
-            delete_template: RangeTemplate {
-                point: format!("DELETE FROM {table_name} WHERE k = $1 RETURNING k,v;"),
-                range: format!("DELETE FROM {table_name} WHERE k >= $1 AND k < $2 RETURNING k,v;"),
-                full: format!("DELETE FROM {table_name} RETURNING k,v"),
-                left_bounded: format!("DELETE FROM {table_name} WHERE k >= $1 RETURNING k,v;"),
-                prefix: format!("DELETE FROM {table_name} WHERE k LIKE $1 RETURNING k,v;"),
-            },
-        }
-    }
-}
-
-/// Templates for the given table name.
-#[derive(Debug, Clone)]
-pub struct SqlTemplateSet {
-    table_name: String,
-    create_table_statement: String,
-    range_template: RangeTemplate,
-    delete_template: RangeTemplate,
-}
-
-impl SqlTemplateSet {
-    /// Generates the sql for batch get.
-    fn generate_batch_get_query(&self, key_len: usize) -> String {
-        let table_name = &self.table_name;
-        let in_clause = generate_in_placeholders(1, key_len).join(", ");
-        format!("SELECT k, v FROM {table_name} WHERE k in ({});", in_clause)
-    }
-
-    /// Generates the sql for batch delete.
-    fn generate_batch_delete_query(&self, key_len: usize) -> String {
-        let table_name = &self.table_name;
-        let in_clause = generate_in_placeholders(1, key_len).join(", ");
-        format!(
-            "DELETE FROM {table_name} WHERE k in ({}) RETURNING k,v;",
-            in_clause
-        )
-    }
-
-    /// Generates the sql for batch upsert.
-    fn generate_batch_upsert_query(&self, kv_len: usize) -> String {
-        let table_name = &self.table_name;
-        let in_placeholders: Vec<String> = (1..=kv_len).map(|i| format!("${}", i)).collect();
-        let in_clause = in_placeholders.join(", ");
-        let mut param_index = kv_len + 1;
-        let mut values_placeholders = Vec::new();
-        for _ in 0..kv_len {
-            values_placeholders.push(format!("(${0}, ${1})", param_index, param_index + 1));
-            param_index += 2;
-        }
-        let values_clause = values_placeholders.join(", ");
-
-        format!(
-            r#"
-    WITH prev AS (
-        SELECT k,v FROM {table_name} WHERE k IN ({in_clause})
-    ), update AS (
-    INSERT INTO {table_name} (k, v) VALUES
-        {values_clause}
-    ON CONFLICT (
-        k
-    ) DO UPDATE SET
-        v = excluded.v
-    )
-
-    SELECT k, v FROM prev;
-    "#
-        )
-    }
-}
 
 /// Query executor for rds. It can execute queries or generate a transaction executor.
 #[async_trait::async_trait]
@@ -289,7 +48,13 @@ pub trait DefaultQueryExecutor: Send + Sync {
 
     fn name() -> &'static str;
 
-    async fn default_query(&self, query: &str, params: &[&Vec<u8>]) -> Result<Vec<KeyValue>>;
+    async fn default_query(&mut self, query: &str, params: &[&Vec<u8>]) -> Result<Vec<KeyValue>>;
+
+    /// Some queries don't need to return any result, such as `DELETE`.
+    async fn default_execute(&mut self, query: &str, params: &[&Vec<u8>]) -> Result<()> {
+        self.default_query(query, params).await?;
+        Ok(())
+    }
 
     async fn txn_executor<'a>(&'a mut self) -> Result<Self::TxnExecutor<'a>>;
 }
@@ -297,7 +62,12 @@ pub trait DefaultQueryExecutor: Send + Sync {
 /// Transaction query executor for rds. It can execute queries in transaction or commit the transaction.
 #[async_trait::async_trait]
 pub trait TxnQueryExecutor<'a>: Send + Sync {
-    async fn txn_query(&self, query: &str, params: &[&Vec<u8>]) -> Result<Vec<KeyValue>>;
+    async fn txn_query(&mut self, query: &str, params: &[&Vec<u8>]) -> Result<Vec<KeyValue>>;
+
+    async fn txn_execute(&mut self, query: &str, params: &[&Vec<u8>]) -> Result<()> {
+        self.txn_query(query, params).await?;
+        Ok(())
+    }
 
     async fn txn_commit(self) -> Result<()>;
 }
@@ -311,15 +81,15 @@ pub trait ExecutorFactory<T: DefaultQueryExecutor>: Send + Sync {
 }
 
 /// Rds backed store for metsrv
-pub struct RdsStore<T, S>
+pub struct RdsStore<T, S, R>
 where
     T: DefaultQueryExecutor + Send + Sync,
     S: ExecutorFactory<T> + Send + Sync,
 {
     max_txn_ops: usize,
-    sql_template_set: SqlTemplateSet,
     txn_retry_count: usize,
     executor_factory: S,
+    sql_template_set: R,
     _phantom: PhantomData<T>,
 }
 
@@ -329,7 +99,7 @@ pub enum RdsQueryExecutor<'a, T: DefaultQueryExecutor + 'a> {
 }
 
 impl<T: DefaultQueryExecutor> RdsQueryExecutor<'_, T> {
-    async fn query(&self, query: &str, params: &Vec<&Vec<u8>>) -> Result<Vec<KeyValue>> {
+    async fn query(&mut self, query: &str, params: &Vec<&Vec<u8>>) -> Result<Vec<KeyValue>> {
         match self {
             Self::Default(executor) => executor.default_query(query, params).await,
             Self::Txn(executor) => executor.txn_query(query, params).await,
@@ -344,42 +114,17 @@ impl<T: DefaultQueryExecutor> RdsQueryExecutor<'_, T> {
     }
 }
 
-impl<T, S> RdsStore<T, S>
-where
-    T: DefaultQueryExecutor + Send + Sync,
-    S: ExecutorFactory<T> + Send + Sync,
-{
+#[async_trait::async_trait]
+pub trait KvQueryExecutor<T: DefaultQueryExecutor> {
     async fn range_with_query_executor(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         req: RangeRequest,
-    ) -> Result<RangeResponse> {
-        let template_type = range_template(&req.key, &req.range_end);
-        let template = self.sql_template_set.range_template.get(template_type);
-        let params = template_type.build_params(req.key, req.range_end);
-        let params_ref = params.iter().collect::<Vec<_>>();
-        // Always add 1 to limit to check if there is more data
-        let query =
-            RangeTemplate::with_limit(template, if req.limit == 0 { 0 } else { req.limit + 1 });
-        let limit = req.limit as usize;
-        debug!("query: {:?}, params: {:?}", query, params);
-        let mut kvs = query_executor.query(&query, &params_ref).await?;
-        if req.keys_only {
-            kvs.iter_mut().for_each(|kv| kv.value = vec![]);
-        }
-        // If limit is 0, we always return all data
-        if limit == 0 || kvs.len() <= limit {
-            return Ok(RangeResponse { kvs, more: false });
-        }
-        // If limit is greater than the number of rows, we remove the last row and set more to true
-        let removed = kvs.pop();
-        debug_assert!(removed.is_some());
-        Ok(RangeResponse { kvs, more: true })
-    }
+    ) -> Result<RangeResponse>;
 
     async fn put_with_query_executor(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         req: PutRequest,
     ) -> Result<PutResponse> {
         let kv = KeyValue {
@@ -407,90 +152,39 @@ where
 
     async fn batch_put_with_query_executor(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         req: BatchPutRequest,
-    ) -> Result<BatchPutResponse> {
-        let mut in_params = Vec::with_capacity(req.kvs.len() * 3);
-        let mut values_params = Vec::with_capacity(req.kvs.len() * 2);
-
-        for kv in &req.kvs {
-            let processed_key = &kv.key;
-            in_params.push(processed_key);
-
-            let processed_value = &kv.value;
-            values_params.push(processed_key);
-            values_params.push(processed_value);
-        }
-        in_params.extend(values_params);
-        let params = in_params.iter().map(|x| x as _).collect::<Vec<_>>();
-        let query = self
-            .sql_template_set
-            .generate_batch_upsert_query(req.kvs.len());
-        let kvs = query_executor.query(&query, &params).await?;
-        if req.prev_kv {
-            Ok(BatchPutResponse { prev_kvs: kvs })
-        } else {
-            Ok(BatchPutResponse::default())
-        }
-    }
+    ) -> Result<BatchPutResponse>;
 
     /// Batch get with certain client. It's needed for a client with transaction.
     async fn batch_get_with_query_executor(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         req: BatchGetRequest,
-    ) -> Result<BatchGetResponse> {
-        if req.keys.is_empty() {
-            return Ok(BatchGetResponse { kvs: vec![] });
-        }
-        let query = self
-            .sql_template_set
-            .generate_batch_get_query(req.keys.len());
-        let params = req.keys.iter().map(|x| x as _).collect::<Vec<_>>();
-        let kvs = query_executor.query(&query, &params).await?;
-        Ok(BatchGetResponse { kvs })
-    }
+    ) -> Result<BatchGetResponse>;
 
     async fn delete_range_with_query_executor(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         req: DeleteRangeRequest,
-    ) -> Result<DeleteRangeResponse> {
-        let template_type = range_template(&req.key, &req.range_end);
-        let template = self.sql_template_set.delete_template.get(template_type);
-        let params = template_type.build_params(req.key, req.range_end);
-        let params_ref = params.iter().map(|x| x as _).collect::<Vec<_>>();
-        let kvs = query_executor.query(template, &params_ref).await?;
-        let mut resp = DeleteRangeResponse::new(kvs.len() as i64);
-        if req.prev_kv {
-            resp.with_prev_kvs(kvs);
-        }
-        Ok(resp)
-    }
+    ) -> Result<DeleteRangeResponse>;
 
     async fn batch_delete_with_query_executor(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         req: BatchDeleteRequest,
-    ) -> Result<BatchDeleteResponse> {
-        if req.keys.is_empty() {
-            return Ok(BatchDeleteResponse::default());
-        }
-        let query = self
-            .sql_template_set
-            .generate_batch_delete_query(req.keys.len());
-        let params = req.keys.iter().map(|x| x as _).collect::<Vec<_>>();
-        let kvs = query_executor.query(&query, &params).await?;
-        if req.prev_kv {
-            Ok(BatchDeleteResponse { prev_kvs: kvs })
-        } else {
-            Ok(BatchDeleteResponse::default())
-        }
-    }
+    ) -> Result<BatchDeleteResponse>;
+}
 
+impl<T, S, R> RdsStore<T, S, R>
+where
+    Self: KvQueryExecutor<T> + Send + Sync,
+    T: DefaultQueryExecutor + Send + Sync,
+    S: ExecutorFactory<T> + Send + Sync,
+{
     async fn execute_txn_cmp(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         cmp: &[Compare],
     ) -> Result<bool> {
         let batch_get_req = BatchGetRequest {
@@ -517,7 +211,7 @@ where
     /// Execute a batch of transaction operations. This function is only used for transactions with the same operation type.
     async fn try_batch_txn(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         txn_ops: &[TxnOp],
     ) -> Result<Option<Vec<TxnOpResponse>>> {
         if !check_txn_ops(txn_ops)? {
@@ -533,7 +227,7 @@ where
 
     async fn handle_batch_delete(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         txn_ops: &[TxnOp],
     ) -> Result<Option<Vec<TxnOpResponse>>> {
         let mut batch_del_req = BatchDeleteRequest {
@@ -568,7 +262,7 @@ where
 
     async fn handle_batch_put(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         txn_ops: &[TxnOp],
     ) -> Result<Option<Vec<TxnOpResponse>>> {
         let mut batch_put_req = BatchPutRequest {
@@ -597,7 +291,7 @@ where
 
     async fn handle_batch_get(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         txn_ops: &[TxnOp],
     ) -> Result<Option<Vec<TxnOpResponse>>> {
         let mut batch_get_req = BatchGetRequest { keys: vec![] };
@@ -636,7 +330,7 @@ where
 
     async fn execute_txn_op(
         &self,
-        query_executor: &RdsQueryExecutor<'_, T>,
+        query_executor: &mut RdsQueryExecutor<'_, T>,
         op: &TxnOp,
     ) -> Result<TxnOpResponse> {
         match op {
@@ -685,7 +379,7 @@ where
 
     async fn txn_inner(&self, txn: &KvTxn) -> Result<KvTxnResponse> {
         let mut default_executor = self.executor_factory.default_executor().await?;
-        let txn_executor = RdsQueryExecutor::Txn(
+        let mut txn_executor = RdsQueryExecutor::Txn(
             self.executor_factory
                 .txn_executor(&mut default_executor)
                 .await?,
@@ -693,26 +387,32 @@ where
         let mut success = true;
         if txn.c_when {
             success = self
-                .execute_txn_cmp(&txn_executor, &txn.req.compare)
+                .execute_txn_cmp(&mut txn_executor, &txn.req.compare)
                 .await?;
         }
         let mut responses = vec![];
         if success && txn.c_then {
-            match self.try_batch_txn(&txn_executor, &txn.req.success).await? {
+            match self
+                .try_batch_txn(&mut txn_executor, &txn.req.success)
+                .await?
+            {
                 Some(res) => responses.extend(res),
                 None => {
                     for txnop in &txn.req.success {
-                        let res = self.execute_txn_op(&txn_executor, txnop).await?;
+                        let res = self.execute_txn_op(&mut txn_executor, txnop).await?;
                         responses.push(res);
                     }
                 }
             }
         } else if !success && txn.c_else {
-            match self.try_batch_txn(&txn_executor, &txn.req.failure).await? {
+            match self
+                .try_batch_txn(&mut txn_executor, &txn.req.failure)
+                .await?
+            {
                 Some(res) => responses.extend(res),
                 None => {
                     for txnop in &txn.req.failure {
-                        let res = self.execute_txn_op(&txn_executor, txnop).await?;
+                        let res = self.execute_txn_op(&mut txn_executor, txnop).await?;
                         responses.push(res);
                     }
                 }
@@ -728,8 +428,10 @@ where
 }
 
 #[async_trait::async_trait]
-impl<T, S> KvBackend for RdsStore<T, S>
+impl<T, S, R> KvBackend for RdsStore<T, S, R>
 where
+    R: 'static,
+    Self: KvQueryExecutor<T> + Send + Sync,
     T: DefaultQueryExecutor + 'static,
     S: ExecutorFactory<T> + 'static,
 {
@@ -743,48 +445,50 @@ where
 
     async fn range(&self, req: RangeRequest) -> Result<RangeResponse> {
         let client = self.executor_factory.default_executor().await?;
-        let query_executor = RdsQueryExecutor::Default(client);
-        self.range_with_query_executor(&query_executor, req).await
+        let mut query_executor = RdsQueryExecutor::Default(client);
+        self.range_with_query_executor(&mut query_executor, req)
+            .await
     }
 
     async fn put(&self, req: PutRequest) -> Result<PutResponse> {
         let client = self.executor_factory.default_executor().await?;
-        let query_executor = RdsQueryExecutor::Default(client);
-        self.put_with_query_executor(&query_executor, req).await
+        let mut query_executor = RdsQueryExecutor::Default(client);
+        self.put_with_query_executor(&mut query_executor, req).await
     }
 
     async fn batch_put(&self, req: BatchPutRequest) -> Result<BatchPutResponse> {
         let client = self.executor_factory.default_executor().await?;
-        let query_executor = RdsQueryExecutor::Default(client);
-        self.batch_put_with_query_executor(&query_executor, req)
+        let mut query_executor = RdsQueryExecutor::Default(client);
+        self.batch_put_with_query_executor(&mut query_executor, req)
             .await
     }
 
     async fn batch_get(&self, req: BatchGetRequest) -> Result<BatchGetResponse> {
         let client = self.executor_factory.default_executor().await?;
-        let query_executor = RdsQueryExecutor::Default(client);
-        self.batch_get_with_query_executor(&query_executor, req)
+        let mut query_executor = RdsQueryExecutor::Default(client);
+        self.batch_get_with_query_executor(&mut query_executor, req)
             .await
     }
 
     async fn delete_range(&self, req: DeleteRangeRequest) -> Result<DeleteRangeResponse> {
         let client = self.executor_factory.default_executor().await?;
-        let query_executor = RdsQueryExecutor::Default(client);
-        self.delete_range_with_query_executor(&query_executor, req)
+        let mut query_executor = RdsQueryExecutor::Default(client);
+        self.delete_range_with_query_executor(&mut query_executor, req)
             .await
     }
 
     async fn batch_delete(&self, req: BatchDeleteRequest) -> Result<BatchDeleteResponse> {
         let client = self.executor_factory.default_executor().await?;
-        let query_executor = RdsQueryExecutor::Default(client);
-        self.batch_delete_with_query_executor(&query_executor, req)
+        let mut query_executor = RdsQueryExecutor::Default(client);
+        self.batch_delete_with_query_executor(&mut query_executor, req)
             .await
     }
 }
 
 #[async_trait::async_trait]
-impl<T, S> TxnService for RdsStore<T, S>
+impl<T, S, R> TxnService for RdsStore<T, S, R>
 where
+    Self: KvQueryExecutor<T> + Send + Sync,
     T: DefaultQueryExecutor + 'static,
     S: ExecutorFactory<T> + 'static,
 {
